Add files via upload
Start project from the radiobutton

Be sure to see..

Please read all the comments of the timeflyer class. It was the major problem I solved doing this project.  
Image animation comments on line 51 of Player 2 class
Motion boundaries on line 35 of Player 4 class
Comments on the window listener in the Frame 1 class
Level stucture comments in the Level 1 class
The comments I directed you to at the top of the Levels 3 and 4 classes
Comments on the use of the rectangle object in the ObjectBlueprint class
The random image switch in the PeopleSprite class